Drop ipaddr as a dependency
For https://gitlab.com/gitlab-org/gitlab/-/issues/376417, we added
ipaddr v1.2.5 as a dependency in
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/115405 in
preparation for Ruby 3.1. There were breaking changes in ipaddr v1.2.3
that came with the upgrade.

ipaddr v1.2.6 ships as a default gem (https://stdgems.org/ipaddr/) in
Ruby 3.3. Having v1.2.5 in the `Gemfile` causes a dependency conflicts
with Bundler since it already activates v1.2.6. To avoid issues with
Ruby 3.3, we can just drop the dependency in `Gemfile`.

Changelog: changed