Show all approvers if user is a member of the project
If the user is a member of the project then it is safe to
show all of the approvers because the user can already
see all of these users on the members page.

If the user is not a member of the project then we fallback
to the previous logic of hiding all approvers.

https://gitlab.com/gitlab-org/gitlab/-/issues/457490