Add instrumentation logger for Ci::RegisterJobService
`Ci::RegisterJobService` sometimes takes a longer time than it should.
However, it's not easy to debug the slowness with the current
instrumentation tools.

In this change, we are adding a new logger, inspired by
the Pipeline Logger, to this service.

Currently, the log condition is 5 seconds. And it's behind an ops
feature flag; `ci_register_job_instrumentation_logger`.