Fix text color of rich text suggestion added line text
Currently when using the rich text editor to add a suggestion, the text
color for added lines is inherited, which makes them very difficult to
read on the dark green background.

This is because the wrong one of the two versions of the diff text was
styled; the one the user edits directly should be transparent and the
one formatted for display should be shown, but the transparent text
class was applied to the wrong one.

Changelog: fixed