Save the LDAP tab pressed in cookie
When sign-in pages have LDAP tabs, preserve the latest tab pressed,
`current_signin_tab`, in the cookie instead of localStorage. This allows
for the value to be preserve on sign out.

In addition, we preserve the value of `preferred_language` which is also
store in the cookie. In this way, the language is not lost on every sign
out.

Changelog: changed