Fix display of CLI broadcast messages with user target roles
Broadcast messages with "Target roles" were not displayed on the CLI.

Why?
The `user_access_level` was always passed in as `nil` into the finder.
So any broadcast message with a non-empty `target_access_levels` checked
if one of its allowed levels was `nil`, which was always false.

Changlog: fixed