Fix flaky test for Merge Requests REST API
Contributes to https://gitlab.com/gitlab-org/gitlab/-/issues/511941

**Problem**

When multiple assignees are provided to the merge request endpoint, then
the assignee field will be populated by one of them. It's not guaranteed
that it will be the first assignee in the list.

**Solution**

Adjust the test to check if any of provided assignees are set in
`assignee` field.