Introduce codeSuggestionsContexts for CurrentUser
This introduces a new field to the CurrentUser GraphQL type.
This new field indicates which additional context categories
are available for the user.

The open_tabs context is available as long as the user has
access to code suggestions. The imports context is toggled
by a Feature Flag.

Changelog: added
EE: true