Improve throttling during user record reassignment
This improves throttling of the placeholder user record
reassignment process. Instead of re-enqueueing the service
as soon as a single db table is unhealthy, we save
the event in a local array, continue to check the next
table, but refer to the local array at the end of the
reassignment process, re-enqueuing the service if the array
is not empty.

Changelog: added