Adds `clientSubscriptionId` to summarize code changes subscription
Adds the `clientSubscriptionId` variable so that when a AI response subscription
is recieved it'll only update the description of the merge request the user
is viewing, and not other merge requests.
It also fixes bugs where other AI features across GitLab might populate the
description if a response is recieved.