Add new usage tables
Project and namespace are in seperate databases
so can't be actual foreign keys. We are sharding
by those columns so they can't be null either
we also want to keep around the namespace and
and project attribution even when namespaces
are deleted. This allows for more auditable
billing records. We certainly never want to
delete a usage record when a project or
runner is deleted because we stil plan
to bill those users retroactively

Add move row to archive table

Add hosted runner monthly usages