Remove unnecessary multiple_todos stub
This fixes a few specs, which had the `multiple_todos: false` stub, when
actually they shouldn't. In all of these specs, entities are receiving a
`MARKED` todo. That means the user is marking them manually via the API
or UI by clicking a button. You can only have one pending `MARKED` todo
per target. So the specs should be working, regardless of the setting of
the feature flag.