Enable specifying a semantic element for the CRUD component container
Currently, the CRUD component renders a `<section>` at its root. This MR
leaves that default in place, but adds a prop to the Vue version and an
option to the ViewComponent version that allows the dev to specify a
different semantic element for the root of the component. This enables
us to avoid some situations where using a `<section>` would be
semantically inappropriate.

Changelog: changed