Move change status part of vulnerability bulk updates to new component
This moves the change status part of the selection summary, which is
used for vulnerability bulk updates on the vulnerability report, to a
new component bulk_change_status. This in preparation of adding a new
action dropdown to the selection summary which will allow multiple
actions, one of which is to change the status.