Use primary DB when authenticating via job token in jobs API
Previously these endpoints used a replica database when authenticating,
which is problematic when using a job token as a new job record may not
exist yet on the replica, causing a 401 error.

To solve this, the primary database will now always be used when
authenticating for these endpoints.

See https://gitlab.com/gitlab-org/gitlab/-/issues/512524.

Changelog: fixed