Update protected tags documentation
The backend logic has checked if a user can create a protected tag when
attempting to delete the protected tag for a long time but the documentation
still states that maintainers can delete them.

This change updates the documentation to ensure that this is correctly
documented.

Related to https://gitlab.com/gitlab-org/gitlab/-/issues/455799
Closes https://gitlab.com/gitlab-org/gitlab/-/work_items/512255

This is 1 piece in a series of fixes for this feature. The frontend currently
only allows maintainers to click the button. We will also be updating the label
for `Allowed to create` to `Allowed to create and delete` as referenced in this
change.