Cache database once per pipeline
Every RSpec CI job uses the `setup_db` to set up a Postgres test
database. This takes about 1m 10s because the database has to be created
and then all migrations have to be applied.

In a pipeline that runs 150 RSpec jobs, this amounts to 3 hours of
work that is needlessly ran 150 times instead of once.

Therefore, this adds a CI job `db:setup pg14` that sets up a database
and makes it available as dump (using `pg_dumpall`) to all RSpec jobs
that test using Postgres 14.

By doing this, and applying the dump in each job with `psql`, we'll
reliably and (hopefully) permanently save about 50 seconds per RSpec
job, or about 3% of total job duration.