Fix epic creation when instance templates are set
When creating an epic, which is creating a work item on a group, we try
to find a description template and set it as content, when the given
description is empty.

This led to an issue for group level work items (for now only epics)
because we assumed that the work item has a project ID.

This fixes also another bug, as the previous code assumed that the
template is in the same project as the issue. However, the template
could come from any project on the instance.

We're still using `project` to find the template, even though `project`
is `nil` in the case of group level work items. This is fine since the
`TemplateFinder` assumes a project, or `nil`.

Changelog: fixed
EE: true