Include SEP jobs with overriding PEPs
This change stops project pipeline being cleared when
overriding pipeline execution policies are in place.

It is no longer necessary because we override the project content
so that it's not loaded in the first place.

This makes it possible for SEP and overriding PEP to be used
at the same time.

Changelog: fixed
EE: true