Refactor dependencies spec as prep for new filter and sort params
Context
-------------------

After a gap analysis of the features available on the group and
project level dependency pages[^1], we have begun work to bring the
pages to parity.

It appears the same [finder][0] is used in the [group][1] and [project][2]
controllers. This means that all we need to do for adding a filter or
sort param is:

1. add the new param to the controller's permitted params[^2]
2. add spec coverage

This change
-------------------

This refactor is prep work for adding the new filters.

It makes the specs written in a way that will hopefully make future
MRs, adding the new filters results in a very readable diff

You can see an example of this spec refactor in use in this subsequent
MR adding a new filter [here][5]

[^1]: There is a table of the analysis on the [epic][4]
[^2]: e.g. the project controller's permitted params are [here][3]

[0]:https://gitlab.com/gitlab-org/gitlab/-/blob/0a5af4787ad597d3f1726d3cd00ee93d16c35e2c/ee/app/finders/sbom/dependencies_finder.rb
[1]:https://gitlab.com/gitlab-org/gitlab/-/blob/0a5af4787ad597d3f1726d3cd00ee93d16c35e2c/ee/app/controllers/groups/dependencies_controller.rb#L104
[2]:https://gitlab.com/gitlab-org/gitlab/-/blob/0a5af4787ad597d3f1726d3cd00ee93d16c35e2c/ee/app/controllers/projects/dependencies_controller.rb#L92
[3]:https://gitlab.com/gitlab-org/gitlab/-/blob/0a5af4787ad597d3f1726d3cd00ee93d16c35e2c/ee/app/controllers/projects/dependencies_controller.rb#L71-81
[4]:https://gitlab.com/groups/gitlab-org/-/epics/15305
[5]:https://gitlab.com/gitlab-org/gitlab/-/merge_requests/179432

---

related to: https://gitlab.com/gitlab-org/gitlab/-/issues/493773
MR: https://gitlab.com/gitlab-org/gitlab/-/merge_requests/179399