Fix Jest Vue 3 nightly jobs
This makes sure the `--vue3` flag is explicitly passed to the nightly
Jest Vue 3 jobs to avoid this failure:

    Expected unset environment variable VUE_VERSION, or VUE_VERSION=2,
    got VUE_VERSION="3".

These nightly jobs are analysed and presented in
https://gitlab-org.gitlab.io/frontend/playground/jest-speed-reporter/vue3.html.

It showed a drop to zero of passing/failing specs under Vue 3 starting
2025-01-18. This the day after
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/175400 merged,
which changed how the `jest_ci.js` script checks its arguments.

Specifically, the `--vue3` flag (or lack of it) must agree with the
`VUE_VERSION` environment variable.