Fix Jest Vue 3 nightly jobs, again
This is another fix for nightly jobs caused by
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/175400.

The test sequencer used was determined entirely by the `vue3` flag.
Before the previous fix
(https://gitlab.com/gitlab-org/gitlab/-/merge_requests/179427), that
flag was *not* being passed, meaning the regular fixture sequencer was
used, even though the the tests were being run under Vue 3.

So, this adds a new `--include-vue3-quarantined` flag to explicitly opt
in to including tests normally quarantined under Vue 3. The test
sequencer is then chosen based on that as well.

Finally, that flag is passed to the nightly `jest* vue3` jobs.