Added backend support for realtime on the merge request dashboard
Adds a new subscription trigger that will trigger whenever a merge request
gets a reviewer assigned, assignee assigned, review requested on a reviewer,
review left by reviewer, a user approves or a user removes their approval.
This subscription update will then be used by the frontend to trigger an update
to each of the affected lists on the homepage.

https://gitlab.com/gitlab-org/gitlab/-/issues/512629