Adds frontend support for realtime updates on the merge request homepage
Adds 1 subscription handler that will wait for any change to any merge request
the current user is an assignee or reviewer of. If there is an update to these
we trigger a refetch of the lists that it affects.
If any merge request gets added to the list we add a green background.
The green background will remain until the next realtime update.
There is also a visibility check, if the user isn't viewing the page and multiple
realtime updates come in we won't clear the green background for
older updates, we only clear it again the user has viewed the page and another
realtime update comes in.

https://gitlab.com/gitlab-org/gitlab/-/issues/512629