Migrate authored todos to GhostUser
Currently, when a user is deleted, we destroy all todos they authored.
It might preferable to change them to the GhostUser, similar to what we
do with Issues, Notes or UserAwards. Just because a user is deleted, it
should not affect my todos.

Hopefully this should fix the problems we are seeing with performance as
well, as the updates are executed in a loop until all todos are
migrated.

Note: In this commit we migrate todos to the GhostUser even if the
hard_delete is set.

See also:
- https://gitlab.com/gitlab-com/gl-infra/production/-/issues/19169
- https://gitlab.com/gitlab-org/gitlab/-/issues/515442

Changelog: changed