Fix CI_RUNNER_TAGS regression
Before changing the taggings association,
calling `job.tag_list` would return two different
object types depending on the state to the tags.
If they were preloaded it would be an array, or if not,
they would be a `TagList` and would have different behavior
when calling `.to_s` on it.

Changelog: fixed