Add NOT NULL constraint to boards sharding key
There are no offenses in .com for records with
no group_id or project_id. But we did find a few records
have have both values present. These records shouldn't
exist, so making the project only as it's safer
than group only since some of the offending records
group_id actually didn't belong to a group but to a user namespace

Changelog: other