Add Cop/ActiveRecordDependent for ee/models
This is related to https://gitlab.com/gitlab-org/gitlab/-/issues/516113
where we discovered that many cops are not being applied to files in the
`ee/` directory.

This MR simply fixes that for `Cop/ActiveRecordDependent` so that we
don't continue adding violations and adds a `rubocop:disable` for all
the existing violations.