Fix UseSqlFunctionForPrimaryKeyLookups specs to avoid View constraint
The specs above are using `projects` to demonstrate the behavior.
When it tries to simulate a column changing its type, it clashes with
the `projects_routes_view` that depend on that particular column.

As the change makes sense only during test, and to avoid clashes with
the recently introduced view, this changes the examples to rely on
`User` instead, for the majority of the cases.