Create qa image for internal release
This is a change for the script to build the QA image.

Internal Release does not use tag, so we need another approach to build
the QA image for it. This change relies on the content of the VERSION
file to determine if it is an internal release, to tag accordingly.

This change also adds new variable BASE_IMAGE_PATH to keep the code DRY.

Ref: https://gitlab.com/gitlab-com/gl-infra/delivery/-/issues/20850