Remove start and due date widget from incident WIT
Removes start and due date widget from
incident work item type because
incidents currently use a legacy implementation
of start/end dates which are not compatible
with the work item widget yet.

So the start and due date widget should
not be visible in the incident create workflow.

Changelog: other