Add a way to present Epics using WorkItem new logic
Roadmap relies on Epic API and we plan to move it to use Work Item API
as a part of https://gitlab.com/gitlab-org/gitlab/-/issues/478011, but
until that the expectation is that there is no significant mismatch of
data between both the APIs especially because other attributes of Epics
are consistent between old and new experience.

To fix that, we're presenting the Epic dates with the `WorkItem` logic.

Related to: https://gitlab.com/gitlab-org/gitlab/-/issues/514066

Changelog: fixed
EE: true