Improve MR vulnerability resolution UX
This change improves the UX when a user uses
the "Resolve with AI" feature on the MR-security
widget.

Currently it can happen that the page needs to do
a hard-fresh once the given vulnerability is
resolved and the added comment, containing the
solutions details, is scrolled into view.

This change decreases the likelihood of this
happening by utilizing a mutation observer.