Split microsoft_applications tables
Migrates data from SystemAccess::MicrosoftApplication to
SystemAccess::GroupMicrosoftApplication. This is to enable the table to
be split between instance application records for instance-wide SAML
sync, and per-group application records for group-based SAML sync.

This way we can shard the data appropriately for Cells.

The code will initially write to both tables during the transition, with
a feature flag to control which table is read from. When the feature
flag is removed, we can remove writing to the legacy table.

Changelog: changed
EE: true