Enable DCR interactions on non-diff MR notes
Formerly, GitLab Duo would only respond to
@mentions when commenting on diff notes that
it created.  In this MR, we enable DCR to respond
to any kind of note on the MR when mentioned.

For non-diff notes, we leverage Duo Chat's merge
request reader tool to provide MR context. We do this
by passing in the MR's iid as additional context.