Add queue_namespace as an attribute for Sidekiq routing rules
This commit adds support for filtering Sidekiq routing rules by
`queue_namespace`, which is a subset of `name`. This is handy if you
want to be able to direct all cronjobs to a separate queue, for
example.

The Developer Experience team is trying to improve the mapping of what
code actually runs when end-to-end tests execute. They saw a lot of
noise in the coverband data showing cronjobs being mapped to unrelated
tests. Being able to run groups of jobs in a different process should
help improve these mappings.

Relates to
https://gitlab.com/gitlab-org/quality/quality-engineering/team-tasks/-/issues/2993

Changelog: added