Add post-merge preloads to predefined_merge_request_variables
This method is slow because of the `MergeRequest#approved?` method.
Especially in EE, there are so many logic and it takes a lot of time.

In the codebase, the preloads of the MR associations are handled but not
in this part. I am adding another set of preloads.

Related to https://gitlab.com/gitlab-org/gitlab/-/issues/515980

Feature flag: `ci_merge_request_variables_preload`