Fix duplicate emails sent for PrAT and GrAT expiring tokens
When sending notifications to inherited members of a group or project,
we sometimes send a duplicate of that email. This is because the finders
may return multiple memebrship records for the same group / project
hierarchy and user. These are usually filtered out by using distinct on
user_id and invite_email , however it's possible for a user to be
invited via email to a parent group and have a non-email direct
membership in a subgroup or project. This will produce multiple Member
records from the relevant Finder classes, and will result in duplicate
emails being sent to the same user.

This change attempts to fix the above case for email notifications.

Changelog: changed