Add filtering by package manager to `Sbom::AggregationsFinder`
Context
----------------------

After a gap analysis of the features available on the group and
project level dependency pages[^1], we have begun work to bring the
pages to parity.

For the group level, sometimes the `Sbom::AggregationsFinder` is
used to fetch the occurrences.

This change
----------------------

1. Adds the `package_managers` key as a valid filter param to the finder
2. Adds a spec coverage for the new filter

[^1]: There is a table of the analysis on the [epic][3]
[^2]: More extensive testing of the filtering (e.g. multiple values,
      nil values, etc.) is left to the finder's spec. Here, we are only
      concerned that the parameters are flowing to the finder as expected

Follow-up work
----------------------

This change only adds the filtering to this finder. Passing the params
from the controller and actually using this filter will be done in
subsequent MRs

---

related to: https://gitlab.com/gitlab-org/gitlab/-/issues/493774
MR: https://gitlab.com/gitlab-org/gitlab/-/merge_requests/179530