Try only highlighting line content
Note:

- the 1px gap beneath the matched collapsible line headers, and
- the highlight doesn't (easily) extend to include the section duration
  on the right.

There's also another subtle issue I don't fully understand where if a
line wraps and is matched, the highlight effect causes the line to
reduce in height slightly, shifting everything below it upwards. It's
likely a bug with what I've tried, but evidently it's not as simple as I
would like.

The second issue can probably be fixed as well with more
code, but following your logic for leaning towards the first approach,
not including the duration in the highlight makes sense because that
isn't searched.