Fix dependency proxy with some object storage providers
A customer reported the Docker client returns `missing or empty
Content-Length header`:

```
$ docker pull gitlab.example.com:443/cloud/dependency_proxy/containers/alpine:latest
Error response from daemon: missing or empty Content-Length header
```

I suspect the issue relates to:

* https://gitlab.com/gitlab-org/gitlab-workhorse/-/issues/264
* https://gitlab.com/gitlab-org/gitlab-workhorse/-/issues/267

When a client requests an object to NetApp StorageGRID with
`Accept-Encoding: gzip`, StorageGRID uses HTTP chunked encoding and
omits `Content-Length`.

As mentioned with `go doc http.Transport.DisableCompression`, the Go
client will automatically add `Accept-Encoding: gzip` if
`Accept-Encoding` is not included.

It seems that for ZIP artifacts, we disabled gzip compression in
https://gitlab.com/gitlab-org/gitlab-workhorse/-/merge_requests/508.

However, we should do the same and leave it to the Docker client to
determine whether gzip encoding can be used. On HEAD requests, the
Docker client omits `Accept-Encoding`, while with GET requests it
includes it. This is good because it means large objects will still be
transferred via gzip, while the small metadata won't necessarily be
compressed.

Changelog: fixed