Fix marking done group wiki comments mention todos
Previously, to-do entries from mentions in group wiki user mentions
couldn't be marked as done in the To-Do List view, due to `group_id`
missing in the Todo record. Also, the header was missing the group name.

This fixes setting the `group_id` attribute for wiki comment/note todos
and the group name is displayed.