Fix bug of missing groups in JobTokenAllowlist
We have a bug in the system when a project maintainer tries to inspect
the CI Job token Allowlist in CI/CD settings and the Allowlist contains
a group that the maintainer has no Guest+ permissions, the group is NOT
displayed in the UI even though it exists

We create a basic minimal Group GQL type called
Types::Ci::JobTokenAccessibleGroupType.
This newly added GQL type does NOT have any authorization check and
only exposes minimal set of fields to display the auth logs and
allowlist tables - id name path fullPath webUrl avatarUrl
We use the newly created minimal Group type JobTokenAccessibleGroupType
instead of the main GroupType in the following 2 places.

 For the groupsAllowlist GQL field in Types::Ci::JobTokenScope
- To see inaccessible projects in Allowlist
 For the target field in Types::Ci::JobTokenScope
-> Ci::JobTokenScope::TargetType -> GroupType (Fix in advance)

[Issue](https://gitlab.com/gitlab-org/gitlab/-/issues/514731)

Changelog: fixed