Reload on type change to incident & add promote_to_incident quickaction
Incidents haven't been migrated to work items yet
and so when we change the type to incident
we want the page to reload. This will render
the legacy incidents view and ensure
everything is working correctly.

To ensure this also works with the new drawer
functionality we browse to the webUrl of the
work item which basically is a reload if the user
is on the detail page, but a browse to the detail
view from the drawer experience.

Additionally we make the /promote_to_incident
quick action available for supported
work items.