Add defense-in-depth specs for markdown includes
Following defense-in-depth specs added in
https://gitlab.com/gitlab-org/gitlab/-/issues/510241 it was noted that
GitLab markdown now includes an additional way to include other files.

This commit adds a spec to ensure that malicious attempts to access a
sensitive file path result in the markdown being rendered as is. If
a vulnerability is introduced in the future that causes the file to be
included, the spec will fail and the vulnerable code will not be merged.

Closes: https://gitlab.com/gitlab-org/gitlab/-/issues/513635