Add a test for conflicting docs directories
With the launch of the Hugo-based docs website,
pages sourced from gitlab are no longer nested in
the "ee" directory in the website build.

This introduces the possibility of conflicting
paths with other project docs, so we need to test
that docs MRs do not add a directory with the same
name as another docs content source project.