Ensure issues has fixed dates
When introducing [autoflow workitems events][bug-mr], we introduced a
[bug][bug-issue] where the legacy issues that weren't created with a due
date weren't possible to update that field. This was happening because
we roll up dates for every work item that's not marked as fixed dates,
and since issues were not setting its dates as fixed on creation, we
endup rolling up null values to the dates on update, emptying the due
date field.

To fix that, we're using the same callback on WorkItems and Legacy
Issues to fixed its dates.

[bug-mr]: https://gitlab.com/gitlab-org/gitlab/-/merge_requests/177277
[bug-issue]: https://gitlab.com/gitlab-org/gitlab/-/issues/516049

Changelog: fixed