Strip trailing newline from rendered Kroki and PlantUML diagram source
A trailing newline is necessary to delineate the ending code fence line;
`SyntaxHighlightFilter` does not consider it to be a bona fide empty
line. This change matches that interpretation, and brings the rich text
editor behavior of Kroki and PlantUML diagrams in line with that of code
blocks and Mermaid diagrams.

Changelog: fixed