Use WikiPage::Meta#find_or_create in GraphQL resolver
This handles a corner case, where years-old wiki pages, which don't have
an associatied WikiPage::Meta record, get one created whenever a GraphQL
resolver tries to find one.

This mirrors behavior of wiki pages on creation and updates, where the
record would also be created if doesn't exists.

Previously, this corner case resulted in views for these wiki pages not
rendering the comments section ("Activity"), because of the missing
record.