Add service for migrating a certificate cluster to the agent
Adds a new service that will be used to perform the first step
of migrating a certificate based Kubernetes cluster to use the
GitLab agent. The service creates an agent, an agent token, and
a migration tracking record. In the future, it will use the
cluster and agent details to install the agent into the linked
cluster.

Also included is a basic frontend and controller action, which
serves as a starting point for frontend development.

Both the service and the frontend changes are controlled by
the `cluster_agent_migrations` feature flag.