Add `MALFORMED_ENTRY_OWNER` error for CODEOWNERS
**Problem**

We currently use Entries with spaces for some errors that are actually
Malformed owners. We need to change this so we convey a clear message
of what's actually happening in the error.

**Solution**

We should add a new error type `MALFORMED_ENTRY_OWNER: Malformed Owners`
to be more specific about the error we are returning

Changelog: added
EE: true