Fix incorrect branch access check
This fixes a bug during merge request pipeline creation permission
checks, in which a ref path or a commit SHA was passed where a branch
name was expected. As a result, the check almost always returned
true (and would almost never hit the intended protection check), causing
the entire pipeline creation to fail.

Changelog: fixed