Update MergeRequest merge_params partial
We can now access the squash options through the merge request. This
change updates the partial to get the values from the MergeRequest
instead of the project.

When the branch_rule_squash_settings feature is turned on this will
allow squash options from branch rules to dictate if the squash checkbox
is enabled in an MR view

This change also moves the issuable_squash_option? helper to the
merge_requests_helpers and renames it to merge_request_squash_option? as
this is never used for Issues.