Source squash option from merge request when squashing
Related to https://gitlab.com/gitlab-org/gitlab/-/issues/503235

Previously project administrators could set a squash option for a
project. Recently we introduced custom squash options per branch.

This change modifies the MergeRequests::SquashService so we check if
squashing is forbidden from the correct source. Instead of checking
directly on the project we check via the merge_request which checks if
any protected branches for the current branch have a squash option
defined and then if none exist we fall back to the project settings.