Handle duplicate WikiPage::Meta record conflicts
This handles duplicate WikiPage::Meta records as a result of reverted
changes in
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/175205.

In general this also handles an unhandled corner case, where we just
raised an error when two Meta records for the same wiki page exist.

Now, the older Meta record is selected as the SSOT, and any events and
todos that might've been associated with it are moved to the older
record. Then, the newer record is removed.