Fix incident layout when new issue look is turned on
When the new issue look is turned on, incidents (which do not use work
items yet) are accidentally currently included in the boolean check to
skip adding legacy issuable right sidebar classes to the page. This MR
adjusts the check to make sure the right sidebar classes are correctly
applied for incident pages.

Changelog: fixed