Update merge request widget serializers
This change fetches the squash options from the merge request instead of
the project which now includes custom options from protected branches.

This also removes the presentation from these expose calls as it's not
required.

I've moved one of the EE specs under the EE namespace as this class is
extended in EE

Related to https://gitlab.com/gitlab-org/gitlab/-/issues/503235