Let the docs link linter register two offenses at once
This improves the `Gitlab/DocumentationLinks/Link` cop to make it able
to register two offenses when a given path lacks the `.md` extension and
does not exist. Previously, a missing extension would prevent the cop
from performing the subsequent check.

This also generally improves how the cop reports offenses by targetting
specific parts of the offending code when possible.