Update glab CLI version and use the notes-text-or-file parameter
The old `release-cli` used to handle the `description` parameter itself.
However, the GitLab CLI (`glab`) did not have this feature, it used to
required either `notes` or `notes-file`. We recently implemented
(https://gitlab.com/gitlab-org/cli/-/issues/7762) a new parameter;
`experimental-notes-text-or-file` to cover this case.

In this change, we start using this new flag.