California is outside of India
The original example says "Can my team in the US use the extra licenses
I purchased for my team in India?" and the answer says "No, because CA
is outside of India". While this is certainly true, it reads more like
this has to do with the laws of California and less to do with the fact
that licenses from one region are being applied to users in a different
region.