Fix housekeeper not requiring filter_identifiers
In https://gitlab.com/gitlab-org/gitlab/-/merge_requests/181094 we
improved the filtering capabilities in Housekeeper and added a new class
FilterIdentifiers. But unfortunately we did not require the class before
using it and now it's breaking Housekeeper in CI (see
https://ops.gitlab.net/gitlab-com/database-team/gitlab-com-database-testing/-/jobs/17251336
for an example).

You can also easily test this yourself by running any keep. I've tested
locally with the example in the README.md and this fixes the problem.