Add simple integration test to just run housekeeper
In https://gitlab.com/gitlab-org/gitlab/-/merge_requests/181613 you can
see it was easy to accidentally completely break housekeeper as there is
no integration test for it. That kind of problem (missing a `require`)
cannot be spotted by running RSpec because the order dependence of tests
might (and did) mask the issue.

Adding an integration test here is an easy way to reduce the risk of this
happening again.