Update move work item service to behave closer to move issue service
Update issue service is running within a transaction. Some
associated data e.g. issues, labels, etc is also copied within
the transaction. So if copying some of this associated data failed
the entire move issue operation fails.

We'll want to move away from that in move work item implementation.
For now we are just implementing the work item service to be at parity.
This would result in less unexpected behaviour before we are ready
to document the changes and implement proper failures handling
mechanisms when outside transaction.