Update redirect handling for help links
The new Docs website uses paths that no
longer include an ee prefix or .html
suffix, and index.md files are now
named _index.md for compatability with
Hugo.

In order for help links to redirect to
the correct page on the website, we need
to adjust the logic that translates a
markdown file path to its webpage counterpart.