Use the pipeline creation meta data rather than project.ci?
When determining if a MR has a pipeline setup, we are changing
to use the new pipeline creation metadata rather than
the project.ci? as the project.ci? will take into account for
auto devops. This was causing issues with the auto merge
having expecting a pipeline when the MR wasn't expecting them
but the project had auto devops on.

MR: gitlab.com/gitlab-org/gitlab/-/merge_requests/181745

Changelog: other