Make it possible for ignore unexpected EOFs in SSL connections
GitLab 17.7 upgraded to OpenSSL 3. With this upgrade, some GitLab
users encountered `unexpected eof while reading` errors with
third-party services, such as NetApp StorageGRID. This error occurs
because OpenSSL 3 requires the server to send a close_notify signal
before shutting down the connection.

This commit adds the ability to disable this check with the
`SSL_IGNORE_UNEXPECTED_EOF` environment variable. Note that this is
not recommended as a default, but we're providing the ability to work
around non-compliant services.

Many discussions have surfaced around the OpenSSL 3 behavior:

* https://github.com/ruby/openssl/pull/730
* https://github.com/redis/redis-rb/issues/1106
* https://bugs.python.org/issue43794
* https://github.com/openssl/openssl/discussions/24810

Relates to https://gitlab.com/gitlab-com/request-for-help/-/issues/2356